[picontrol_uinput.c] Updated comment to suggest better type
At least I think it's better..

+++ src/backend/picontrol_uinput.c
@@ -14,7 +14,7 @@ static const pictrl_key_range valid_key_ranges[] = {
 Index ("key") = ascii char
 Entry ("value") = keyscan combination to produce the ascii
 
-Ex. pictrl_ascii_to_event_codes[(unsigned int)"H" = 0x48] = [KEY_LEFTSHIFT, KEY_H]
+Ex. pictrl_ascii_to_event_codes[(size_t)"H" = 0x48] = [KEY_LEFTSHIFT, KEY_H]
 */
 static const pictrl_key_combo pictrl_ascii_to_event_codes[] = {
     // TODO: Make these repetitive ones a macro or something?
